move server context to common
